+++ src/main/java/com/example/medicalqnaclient/page/core/ViewController.java
@@ -1,6 +1,6 @@
 package com.example.medicalqnaclient.page.core;
 
-import com.example.medicalqnaclient.page.event.Publisher;
+import com.example.medicalqnaclient.page.event.publishers.QnAPublisher;
 import com.example.medicalqnaclient.user.mediator.UserMediator;
 import javafx.scene.layout.Pane;
 
@@ -12,10 +12,10 @@ import javafx.scene.layout.Pane;
  * 3. 이벤트를 발행
  */
 public abstract class ViewController {
-    protected final Publisher publisher;
+    protected final QnAPublisher publisher;
     private final UserMediator mediator;
 
-    public ViewController(Publisher publisher, UserMediator mediator) {
+    public ViewController(QnAPublisher publisher, UserMediator mediator) {
         this.publisher = publisher;
         this.mediator = mediator;
     }
