+++ src/main/java/com/example/medicalqnaclient/page/event/listeners/QnAListener.java
@@ -1,16 +1,18 @@
-package com.example.medicalqnaclient.page.event;
+package com.example.medicalqnaclient.page.event.listeners;
 
 import com.example.medicalqnaclient.page.core.PageManager;
+import com.example.medicalqnaclient.page.event.events.GoHomeEvent;
+import com.example.medicalqnaclient.page.manager.application.qna.QnAPageManager;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
 import static com.example.medicalqnaclient.page.core.PageType.HOME_PAGE;
 
 @Component
-public class Listener {
+public class QnAListener {
     private final PageManager pageManager;
 
-    public Listener(PageManager pageManager) {
+    public QnAListener(QnAPageManager pageManager) {
         this.pageManager = pageManager;
     }
